+++ telco_si/target/debug/telco_si.d
@@ -1,1 +1,0 @@
-/home/dali/WORK/utils/telco_si/target/debug/telco_si: /home/dali/WORK/utils/telco_si/src/main.rs